Add service set member score
